daemon-mgr/requirements.txt
pyyaml
pytest  # for testing
